Added the curve25519-donna code to provide support for all platforms.
See http://code.google.com/p/curve25519-donna/ for the code.

The README indicates it is licensed under the BSD license, but provides
neither a copy of the license, nor a copyright date or author, so I'm
not sure how to credit it other than with the URL above.

Snapshot taken 2011-07-20